+++ crates/matrix-sdk-crypto/src/identities/manager.rs
@@ -20,6 +20,7 @@ use std::{
 };
 
 use futures_util::future::join_all;
+use itertools::Itertools;
 use matrix_sdk_common::{
     executor::spawn,
     locks::Mutex,
@@ -39,7 +40,10 @@ use crate::{
     },
     olm::PrivateCrossSigningIdentity,
     requests::KeysQueryRequest,
-    store::{Changes, DeviceChanges, IdentityChanges, Result as StoreResult, Store},
+    store::{
+        caches::SequenceNumber, Changes, DeviceChanges, IdentityChanges, Result as StoreResult,
+        Store,
+    },
     types::{CrossSigningKey, DeviceKeys, MasterPubkey, SelfSigningPubkey, UserSigningPubkey},
     utilities::FailuresCache,
     LocalTrust, SignatureError,
@@ -128,7 +132,7 @@ pub(crate) struct IdentityManager {
     store: Store,
 
     /// Details of the current "in-flight" key query request, if any
-    keys_query_request_details: Arc<Mutex<KeysQueryRequestDetails>>,
+    keys_query_request_details: Arc<Mutex<Option<KeysQueryRequestDetails>>>,
 }
 
 /// Details of an in-flight key query request
@@ -136,7 +140,7 @@ pub(crate) struct IdentityManager {
 struct KeysQueryRequestDetails {
     /// The sequence number, to be passed to
     /// `Store.mark_tracked_users_as_up_to_date`.
-    sequence_number: i64,
+    sequence_number: SequenceNumber,
 
     /// A single batch of queries returned by the Store is broken up into one or
     /// more actual KeysQueryRequests, each with their own request id. We
@@ -149,7 +153,7 @@ impl IdentityManager {
 
     pub fn new(user_id: Arc<UserId>, device_id: Arc<DeviceId>, store: Store) -> Self {
         let keys_query_listener = KeysQueryListener::new(store.clone());
-        let keys_query_request_details = Mutex::new(KeysQueryRequestDetails::default());
+        let keys_query_request_details = Mutex::new(None);
 
         IdentityManager {
             user_id,
@@ -223,11 +227,19 @@ impl IdentityManager {
         // if this request is one of those we expected to be in flight, pass the
         // sequence number back to the store so that it can mark devices up to
         // date
-        let (sequence_number, removed) = {
+        let sequence_number = {
             let mut request_details = self.keys_query_request_details.lock().await;
-            (request_details.sequence_number, request_details.request_ids.remove(request_id))
+
+            request_details.as_mut().and_then(|details| {
+                if details.request_ids.remove(request_id) {
+                    Some(details.sequence_number)
+                } else {
+                    None
+                }
+            })
         };
-        if removed {
+
+        if let Some(sequence_number) = sequence_number {
             self.store
                 .mark_tracked_users_as_up_to_date(
                     response.device_keys.keys().map(Deref::deref),
@@ -686,7 +698,7 @@ impl IdentityManager {
     ///
     /// # Returns
     ///
-    /// A list of pairs of `(request_id, request)`
+    /// A map of a request ID to the `/keys/query` request.
     ///
     /// The response of a successful key query requests needs to be passed to
     /// the [`OlmMachine`] with the [`receive_keys_query_response`].
@@ -695,9 +707,9 @@ impl IdentityManager {
     /// [`receive_keys_query_response`]: #method.receive_keys_query_response
     pub async fn users_for_key_query(
         &self,
-    ) -> StoreResult<Vec<(OwnedTransactionId, KeysQueryRequest)>> {
-        // forget about any previous key queries in flight
-        *(self.keys_query_request_details.lock().await) = KeysQueryRequestDetails::default();
+    ) -> StoreResult<BTreeMap<OwnedTransactionId, KeysQueryRequest>> {
+        // Forget about any previous key queries in flight.
+        *self.keys_query_request_details.lock().await = None;
 
         let (users, sequence_number) = self.store.users_for_key_query().await?;
 
@@ -715,25 +727,44 @@ impl IdentityManager {
             };
 
         if users.is_empty() {
-            return Ok(Vec::new());
+            Ok(BTreeMap::new())
+        } else {
+            // Let's remove users that are part of the `FailuresCache`. The cache, which is
+            // a TTL cache, remembers users for which a previous `/key/query` request has
+            // failed. We don't retry a `/keys/query` for such users for a
+            // certain amount of time.
+            let users = users.into_iter().filter(|u| !self.failures.contains(u.server_name()));
+
+            // We don't want to create a single `/keys/query` request with an infinite
+            // amount of users. Some servers will likely bail out after a
+            // certain amount of users and the responses will be large. In the
+            // case of a transmission error, we'll have to retransmit the large
+            // response.
+            //
+            // Convert the set of users into multiple /keys/query requests.
+            let requests: BTreeMap<_, _> = users
+                .chunks(Self::MAX_KEY_QUERY_USERS)
+                .into_iter()
+                .map(|user_chunk| {
+                    let request_id = TransactionId::new();
+                    let request = KeysQueryRequest::new(user_chunk);
+
+                    debug!(?request_id, users = ?request.device_keys.keys(), "Created a /keys/query request");
+
+                    (request_id, request)
+                })
+                .collect();
+
+            // Collect the request IDs, these will be used later in the
+            // `receive_keys_query_response()` method to figure out if the user can be
+            // marked as up-to-date/non-dirty.
+            let request_ids = requests.keys().cloned().collect();
+            let request_details = KeysQueryRequestDetails { sequence_number, request_ids };
+
+            *self.keys_query_request_details.lock().await = Some(request_details);
+
+            Ok(requests)
         }
-
-        let mut result: Vec<(OwnedTransactionId, KeysQueryRequest)> = Vec::new();
-        let mut request_details =
-            KeysQueryRequestDetails { sequence_number, request_ids: HashSet::new() };
-
-        let filtered_users: Vec<OwnedUserId> =
-            users.into_iter().filter(|u| !self.failures.contains(u.server_name())).collect();
-
-        for c in filtered_users.chunks(Self::MAX_KEY_QUERY_USERS) {
-            let request_id = TransactionId::new();
-            let req = KeysQueryRequest::new(c.iter().map(|u| (u.clone(), Vec::new())).collect());
-            debug!(?request_id, users = ?c, "Created keys query request");
-            result.push((request_id.clone(), req));
-            request_details.request_ids.insert(request_id);
-        }
-        *(self.keys_query_request_details.lock().await) = request_details;
-        Ok(result)
     }
 
     /// Receive the list of users that contained changed devices from the
@@ -1066,6 +1097,7 @@ pub(crate) mod tests {
 
         let listener = manager.listen_for_received_queries();
 
+        #[allow(unknown_lints, clippy::redundant_async_block)] // false positive
         let task = tokio::task::spawn(async move { listener.wait(Duration::from_secs(10)).await });
 
         manager
@@ -1152,7 +1184,7 @@ pub(crate) mod tests {
         manager.update_tracked_users([alice]).await.unwrap();
 
         // alice should be in the list of key queries
-        let (reqid, req) = manager.users_for_key_query().await.unwrap().pop().unwrap();
+        let (reqid, req) = manager.users_for_key_query().await.unwrap().pop_first().unwrap();
         assert!(req.device_keys.contains_key(alice));
 
         // another invalidation turns up
@@ -1162,7 +1194,7 @@ pub(crate) mod tests {
         manager.receive_keys_query_response(&reqid, &other_key_query()).await.unwrap();
 
         // alice should *still* be in the list of key queries
-        let (reqid, req) = manager.users_for_key_query().await.unwrap().pop().unwrap();
+        let (reqid, req) = manager.users_for_key_query().await.unwrap().pop_first().unwrap();
         assert!(req.device_keys.contains_key(alice));
 
         // another key query response
@@ -1188,7 +1220,7 @@ pub(crate) mod tests {
             manager.store.tracked_users().await.unwrap().contains(alice),
             "Alice is tracked after being marked as tracked"
         );
-        let (reqid, req) = manager.users_for_key_query().await.unwrap().pop().unwrap();
+        let (reqid, req) = manager.users_for_key_query().await.unwrap().pop_first().unwrap();
         assert!(req.device_keys.contains_key(alice));
 
         // a failure should stop us querying for the user's keys.