+++ crates/matrix-sdk-crypto/src/store/caches.rs
@@ -17,11 +17,15 @@
 //! Note: You'll only be interested in these if you are implementing a custom
 //! `CryptoStore`.
 
-use std::{collections::HashMap, sync::Arc};
+use std::{
+    collections::{HashMap, HashSet},
+    sync::Arc,
+};
 
 use dashmap::DashMap;
 use matrix_sdk_common::locks::Mutex;
 use ruma::{DeviceId, OwnedDeviceId, OwnedRoomId, OwnedUserId, RoomId, UserId};
+use tracing::{field::debug, instrument, trace, Span};
 
 use crate::{
     identities::ReadOnlyDevice,
@@ -163,16 +167,125 @@ impl DeviceStore {
     }
 }
 
+/// A numeric type that can represent an infinite ordered sequence.
+///
+/// It uses wrapping arithmetic to make sure we never run out of numbers. (2**64
+/// should be enough for anyone, but it's easy enough just to make it wrap.)
+//
+/// Internally it uses a *signed* counter so that we can compare values via a
+/// subtraction. For example, suppose we've just overflowed from i64::MAX to
+/// i64::MIN. (i64::MAX.wrapping_sub(i64::MIN)) is -1, which tells us that
+/// i64::MAX comes before i64::MIN in the sequence.
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
+pub(crate) struct SequenceNumber(i64);
+
+impl PartialOrd for SequenceNumber {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.0.wrapping_sub(other.0).cmp(&0))
+    }
+}
+
+impl Ord for SequenceNumber {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.0.wrapping_sub(other.0).cmp(&0)
+    }
+}
+
+impl SequenceNumber {
+    fn increment(&mut self) {
+        self.0 = self.0.wrapping_add(1)
+    }
+
+    fn previous(&self) -> Self {
+        Self(self.0.wrapping_sub(1))
+    }
+}
+
+/// Record of the users that are waiting for a /keys/query.
+///
+/// To avoid races, we maintain a sequence number which is updated each time we
+/// receive an invalidation notification. We also record the sequence number at
+/// which each user was last invalidated. Then, we attach the current sequence
+/// number to each `/keys/query` request, and when we get the response we can
+/// tell if any users have been invalidated more recently than that request.
+#[derive(Debug)]
+pub(super) struct UsersForKeyQuery {
+    /// The sequence number we will assign to the next addition to user_map
+    next_sequence_number: SequenceNumber,
+
+    /// The users pending a lookup, together with the sequence number at which
+    /// they were added to the list
+    user_map: HashMap<OwnedUserId, SequenceNumber>,
+}
+
+impl UsersForKeyQuery {
+    /// Create a new, empty, `UsersForKeyQueryCache`
+    pub(super) fn new() -> Self {
+        UsersForKeyQuery { next_sequence_number: Default::default(), user_map: Default::default() }
+    }
+
+    /// Record a new user that requires a key query
+    pub(super) fn insert_user(&mut self, user: &UserId) {
+        let seq = self.next_sequence_number;
+
+        trace!(?user, sequence_number = ?seq, "Flagging user for key query");
+
+        self.user_map.insert(user.to_owned(), seq);
+        self.next_sequence_number.increment();
+    }
+
+    /// Record that a user has received an update with the given sequence
+    /// number.
+    ///
+    /// If the sequence number is newer than the oldest invalidation for this
+    /// user, it is removed from the list of those needing an update.
+    ///
+    /// Returns true if the user is now up-to-date, else false
+    #[instrument(level = "trace", skip(self), fields(invalidation_sequence))]
+    pub(super) fn maybe_remove_user(
+        &mut self,
+        user: &UserId,
+        query_sequence: SequenceNumber,
+    ) -> bool {
+        let last_invalidation = self.user_map.get(user).copied();
+
+        if let Some(last_invalidation) = last_invalidation {
+            Span::current().record("invalidation_sequence", debug(last_invalidation));
+
+            if last_invalidation > query_sequence {
+                trace!("User invalidated since this query started: still not up-to-date");
+                false
+            } else {
+                trace!("User now up-to-date");
+                self.user_map.remove(user);
+                true
+            }
+        } else {
+            trace!("User already up-to-date, nothing to do");
+            true
+        }
+    }
+
+    /// Fetch the list of users waiting for a key query, and the current
+    /// sequence number
+    pub(super) fn users_for_key_query(&self) -> (HashSet<OwnedUserId>, SequenceNumber) {
+        // we return the sequence number of the last invalidation
+        let sequence_number = self.next_sequence_number.previous();
+        (self.user_map.keys().cloned().collect(), sequence_number)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use matrix_sdk_test::async_test;
+    use proptest::prelude::*;
     use ruma::room_id;
     use vodozemac::{Curve25519PublicKey, Ed25519PublicKey};
 
+    use super::{DeviceStore, GroupSessionStore, SequenceNumber, SessionStore};
     use crate::{
         identities::device::testing::get_device,
         olm::{tests::get_account_and_session, InboundGroupSession},
-        store::caches::{DeviceStore, GroupSessionStore, SessionStore},
     };
 
     #[async_test]
@@ -262,5 +375,35 @@ mod tests {
 
         let loaded_device = store.get(device.user_id(), device.device_id());
         assert!(loaded_device.is_none());
+    }
+
+    #[test]
+    fn sequence_at_boundary() {
+        let first = SequenceNumber(i64::MAX);
+        let second = SequenceNumber(first.0.wrapping_add(1));
+        let third = SequenceNumber(first.0.wrapping_sub(1));
+
+        assert!(second > first);
+        assert!(first < second);
+        assert!(third < first);
+        assert!(first > third);
+        assert!(second > third);
+        assert!(third < second);
+    }
+
+    proptest! {
+        #[test]
+        fn partial_eq_sequence_number(sequence in i64::MIN..i64::MAX) {
+            let first = SequenceNumber(sequence);
+            let second = SequenceNumber(first.0.wrapping_add(1));
+            let third = SequenceNumber(first.0.wrapping_sub(1));
+
+            assert!(second > first);
+            assert!(first < second);
+            assert!(third < first);
+            assert!(first > third);
+            assert!(second > third);
+            assert!(third < second);
+        }
     }
 }